+++ .prettierignore
@@ -2,3 +2,4 @@
 /dist
 /coverage
 /.nx/cache
+pnpm-lock.yaml
