+++ apps/chameleon/app/(editor)/editor/[postId]/page.tsx
@@ -1,5 +1,5 @@
 import { notFound, redirect } from 'next/navigation'
-import { db } from '@chameleon/data-access/db'
+import { db } from '@chameleon/db'
 import { Post, User } from '@prisma/client'
 
 import { authOptions } from '@/lib/auth'
