+++ apps/chameleon/lib/auth.ts
@@ -1,5 +1,5 @@
 import { PrismaAdapter } from '@auth/prisma-adapter'
-import { db } from '@chameleon/data-access/db'
+import { db } from '@chameleon/db'
 import { NextAuthOptions } from 'next-auth'
 import EmailProvider from 'next-auth/providers/email'
 import GitHubProvider from 'next-auth/providers/github'
