+++ data-access/db/README.md
@@ -1,7 +1,7 @@
-# data-access/db
+# db
 
 This library was generated with [Nx](https://nx.dev).
 
 ## Running unit tests
 
-Run `nx test data-access/db` to execute the unit tests via [Jest](https://jestjs.io).
+Run `nx test db` to execute the unit tests via [Jest](https://jestjs.io).
