+++ data-access/db/project.json
@@ -1,5 +1,5 @@
 {
-  "name": "data-access/db",
+  "name": "db",
   "$schema": "../../node_modules/nx/schemas/project-schema.json",
   "sourceRoot": "data-access/db/src",
   "projectType": "library",
