+++ package.json
@@ -83,7 +83,6 @@
     "react-resizable-panels": "^1.0.9",
     "react-textarea-autosize": "^8.4.1",
     "resend": "^3.1.0",
-    "sharp": "^0.33.2",
     "shiki": "^0.11.1",
     "simplex-noise": "^4.0.1",
     "sonner": "^1.3.1",
